+++ Package.swift
@@ -15,7 +15,10 @@ let package = Package(
         .package(url: "https://github.com/swiftlang/swift-docc-plugin", from: "1.3.0")
     ],
     targets: [
-        .target(name: "WikilayerClient"),
+        .target(
+            name: "WikilayerClient",
+            resources: [.process("Resources")]
+        ),
         .testTarget(name: "WikilayerClientTests", dependencies: ["WikilayerClient"])
     ]
 )